+++ test/functorch/test_aotdispatch.py
@@ -6011,7 +6011,7 @@ class TestAOTModuleSimplified(AOTTestCase):
             aot_module_simplified(MockModule(), (fake_x,), nop)
 
     def test_aot_test_subclasses_with_tensor_factories(self):
-        from torch.testing._internal.test_subclasses import SubclassWithTensorFactory
+        from torch.testing._internal.common_subclass import SubclassWithTensorFactory
 
         inp = SubclassWithTensorFactory(torch.zeros(3, 5))
 
@@ -6020,7 +6020,7 @@ class TestAOTModuleSimplified(AOTTestCase):
 
         ref_out = fn(inp)
         out = torch.compile(fn, backend="aot_eager", fullgraph=True)(inp)
-        # self.assertEqual(ref_out, out)
+        self.assertEqual(ref_out, out)
 
 
 # entries in here don't work and need to be fixed.
